requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pytesseract
pillow
python-multipart
reportlab
itsdangerous
requests
bson